requirements.txt
# -------------------------
# FastAPI server
# -------------------------
fastapi
uvicorn[standard]

# -------------------------
# HTTP requests
# -------------------------
requests

# -------------------------
# OpenAI API
# -------------------------
openai

# -------------------------
# Background workers
# -------------------------
celery
redis

# -------------------------
# environment config
# -------------------------
python-dotenv

# -------------------------
# validation
# -------------------------
pydantic

# -------------------------
# report generation
# -------------------------
reportlab